requirements.txt
flask==3.1.0
youtube-transcript-api==1.0.3
python-dotenv==1.0.0
cryptography==42.0.5 